Restructured so that common files are apparent.
