+++ src/common_R_interface/r2q.c
@@ -1,19 +1,14 @@
-/*
- * R server for Q
- */
+/*-----------------------------------------------*/
+/*  File: r2q.c                                  */
+/*  Overview: common code for R -> Q interface   */
+/*-----------------------------------------------*/
+
 /*
  * The public interface used from Q.
  * https://cran.r-project.org/doc/manuals/r-release/R-ints.pdf
  * https://cran.r-project.org/doc/manuals/r-release/R-exts.html
  */
 
-K ropen(K x);
-K rclose(K x);
-K rexec(K x);
-K rget(K x);
-K rset(K x,K y);
-
-ZK rcmd(int type,K x);
 ZK klogicv(J len, int *val);
 ZK klogica(J len, int rank, int *shape, int *val);
 ZK kintv(J len, int *val);
@@ -22,9 +17,6 @@ ZK klonga(J len, int rank, int *shape, J*val);
 ZK kdoublev(J len, double *val);
 ZK kdoublea(J len, int rank, int *shape, double *val);
 ZK atom_value_dict(J len, K v, SEXP keys);
-
-__thread int ROPEN=-1; // initialise thread-local. Will fail in other threads. Ideally need to check if on q main thread.
-__thread int RLOAD=0;
 
 /*
  * convert R SEXP into K object.
@@ -53,7 +45,6 @@ ZK from_raw_robject(SEXP);
 ZK from_nyi_robject(SEXP);
 ZK from_frame_robject(SEXP);
 ZK from_factor_robject(SEXP);
-
 
 /*
  * Functions to derive month count since kdb epoch from day count
@@ -109,7 +100,41 @@ Rboolean isUnit(const char *units_, SEXP s){
   return FALSE;
 }
 
-ZK from_any_robject(SEXP sxp){
+/*
+ * Utility functions to handle attribute
+ */
+
+//extern ZK from_pairlist_robject(SEXP sxp);
+
+/* add attribute */
+ZK addattR (K x,SEXP att) {
+  // attrs are pairlists: LISTSXP
+  K u = from_pairlist_robject(att);
+  return knk(2,u,x);
+}
+
+/* add attribute if any */
+ZK attR(K x,SEXP sxp) {
+  SEXP att = ATTRIB(sxp);
+  if (isNull(att))
+    return x;
+  return addattR(x,att);
+}
+
+ZK atom_value_dict(J len, K v, SEXP keys){
+  K k= ktn(KS, len);
+  for(J i= 0; i < len; i++) {
+    const char *keyName= CHAR(STRING_ELT(keys, i));
+    kS(k)[i]= ss((S) keyName);
+  }
+  return xD(k,v);
+}
+
+/*
+ * Conversopn from R to Q
+ */
+
+ ZK from_any_robject(SEXP sxp){
   if(isClass("data.frame", sxp))
     return from_frame_robject(sxp);
   if(isClass("factor", sxp))
@@ -184,30 +209,6 @@ ZK from_any_robject(SEXP sxp){
   return result;
 }
 
-/* add attribute */
-ZK addattR (K x,SEXP att) {
-  // attrs are pairlists: LISTSXP
-  K u = from_pairlist_robject(att);
-  return knk(2,u,x);
-}
-
-/* add attribute if any */
-ZK attR(K x,SEXP sxp) {
-  SEXP att = ATTRIB(sxp);
-  if (isNull(att))
-    return x;
-  return addattR(x,att);
-}
-
-ZK atom_value_dict(J len, K v, SEXP keys){
-  K k= ktn(KS, len);
-  for(J i= 0; i < len; i++) {
-    const char *keyName= CHAR(STRING_ELT(keys, i));
-    kS(k)[i]= ss((S) keyName);
-  }
-  return xD(k,v);
-}
-
 ZK error_broken_robject(SEXP sxp) {
   return krr("Broken R object.");
 }
@@ -247,6 +248,7 @@ ZK from_raw_robject(SEXP sxp) {
   DO(xn,kG(x)[i]=RAW(sxp)[i])
   return x;
 }
+
 
 ZK from_date_robject(SEXP sxp) {
   K x;
@@ -337,7 +339,7 @@ ZK from_days_robject(SEXP sxp){
   return x;
 }
 
-//Wrapper function of difftime
+/* Wrapper function of difftime */
 ZK from_difftime_robject(SEXP sxp){
   if(isUnit("secs", sxp) || isUnit("mins", sxp))
     return from_second_or_minute_robject(sxp);
@@ -347,8 +349,10 @@ ZK from_difftime_robject(SEXP sxp){
     return from_nyi_robject(sxp);
 }
 
-// NULL in R(R_NilValue): often used as generic zero length vector
-// NULL objects cannot have attributes and attempting to assign one by attr gives an error
+/*
+ * NULL in R(R_NilValue): often used as generic zero length vector
+ * NULL objects cannot have attributes and attempting to assign one by attr gives an error
+ */
 ZK from_null_robject(SEXP sxp) {
   return knk(0);
 }
@@ -356,17 +360,6 @@ ZK from_null_robject(SEXP sxp) {
 ZK from_symbol_robject(SEXP sxp) {
   const char* t = CHAR(PRINTNAME(sxp));
   K x = ks((S)t);
-  return attR(x,sxp);
-}
-
-ZK from_pairlist_robject(SEXP sxp) {
-  K x = ktn(0,2*length(sxp));
-  SEXP s = sxp;J i;
-  for(i=0;i<x->n;i+=2) {
-    kK(x)[i] = from_any_robject(CAR(s));
-    kK(x)[i+1] = from_any_robject(TAG(s));
-    s=CDR(s);
-  }
   return attR(x,sxp);
 }
 
@@ -517,30 +510,6 @@ ZK from_vector_robject(SEXP sxp) {
 }
 
 /*
- * various utilities
- */
-
-/* get k string or symbol name */
-static char * getkstring(K x) {
-  char *s=NULL;
-  int len;
-  switch (xt) {
-    case -KC :
-      s = calloc(2,1); s[0] = xg;
-      break;
-    case KC :
-      s = calloc(1+xn,1); memmove(s, xG, xn);
-      break;
-    case -KS : // TODO: xs is already 0 terminated and fixed. can just return xs
-      len = 1+strlen(xs);
-      s = calloc(len,1); memmove(s, xs, len); break;
-    default :
-      krr("invalid name");
-  }
-  return s;
-}
-
-/*
  * convert R arrays to K lists
  * done for boolean, int, double
  */
@@ -659,7 +628,9 @@ ZK kdoublea(J len, int rank, int *shape, double *val) {
       x = kdoublev(len,val);
       break;
     case 2 :
-      r = shape[0]; c = shape[1]; x = knk(0);
+      r = shape[0];
+      c = shape[1];
+      x = knk(0);
       for (i=0;i<r;i++) {
         y = ktn(KF,c);
         for (j=0;j<c;j++)
@@ -668,137 +639,12 @@ ZK kdoublea(J len, int rank, int *shape, double *val) {
       };
       break;
     default :
-      k = rank-1; r = shape[k]; c = len/r; x = knk(0);
+      k = rank-1;
+      r = shape[k];
+      c = len/r;
+      x = knk(0);
       for (i=0;i<r;i++)
         x = jk(&x,kdoublea(c,k,shape,val+c*i));
   }
   return x;
 }
-
-/*
- * The public interface used from Q.
- */
-static I spair[2];
-void* pingthread;
-
-V* pingmain(V* v){
-  while(1){
-    nanosleep(&(struct timespec){.tv_sec=0,.tv_nsec=1000000}, NULL);
-    send(spair[1], "M", 1, 0);
-  }
-}
-
-K processR(I d){
-  char buf[1024];
-  /*MSG_DONTWAIT - set in sd1(-h,...) */
-  while(0 < recv(d, buf, sizeof(buf), 0))
-    ;
-  R_ProcessEvents();
-  return (K)0;
-}
-
-/*
- * ropen argument is empty, 0 or 1
- * empty,0   --slave (R is quietest)
- * 1         --verbose
- */
-K ropen(K x) {
-  if(!RLOAD) return krr("main thread only");
-  if (ROPEN >= 0) return ki(ROPEN);
-  int s,mode=0;	char *argv[] = {"R","--slave"};
-  if (x && (-KI ==x->t || -KJ ==x->t)) mode=(x->t==-KI?x->i:x->j)!=0;
-  if (mode) argv[1] = "--verbose";
-  int argc = sizeof(argv)/sizeof(argv[0]);
-  s=Rf_initEmbeddedR(argc, argv);
-  if (s<0) return krr("open failed");
-  if(dumb_socketpair(spair, 1) == -1)
-    return krr("Init failed for socketpair");
-  sd1(-spair[0], &processR);
-  #ifndef WIN32
-    pthread_t t;
-    if(pthread_create(&t, NULL, pingmain, NULL))
-      R krr("poller_thread");
-    pingthread= &t;
-  #else
-    if(_beginthreadex(0,0,pingmain,NULL,0,0)==-1)
-   	R krr("poller_thread");
-  #endif
-  ROPEN=mode;
-  return ki(ROPEN);
-}
-
-// note that embedded R can be initialised once. No open/close/open supported 
-// http://r.789695.n4.nabble.com/Terminating-and-restarting-an-embedded-R-instance-possible-td4641823.html
-K rclose(K x){R NULL;}
-K rexec(K x) { return rcmd(0,x); }
-K rget(K x) { return rcmd(1,x); }
-
-static char* ParseError[5]={"null","ok","incomplete","error","eof"};
-
-
-K rcmd(int type,K x) {
-  if(!RLOAD) return krr("main thread only");
-  if (ROPEN < 0) ropen(NULL);
-  SEXP e, p, r, xp;
-  char rerr[256];extern char	R_ParseErrorMsg[256];
-  int error;
-  ParseStatus status;
-  if(abs(x->t)==KS)
-    e=from_symbol_kobject(x);
-  else if(abs(x->t)==KC)
-    e=from_string_kobject(x);
-  else
-    return krr("type");
-  PROTECT(e);
-  PROTECT(p=R_ParseVector(e, 1, &status, R_NilValue));
-  if (status != PARSE_OK) {
-    UNPROTECT(2);
-    snprintf(rerr,sizeof(rerr),"%s: %s",ParseError[status], R_ParseErrorMsg);
-    return krr(rerr);
-  }
-  PROTECT(xp=VECTOR_ELT(p, 0));
-  r=R_tryEvalSilent(xp, R_GlobalEnv, &error);
-  UNPROTECT(3);
-  R_ProcessEvents();
-  if (error) {
-    snprintf(rerr,sizeof(rerr),"eval error: %s",R_curErrorBuf());
-    return krr(rerr);
-  }
-  if (type==1)
-    return from_any_robject(r);
-  return (K)0;
-}
-
-K rset(K x,K y) {
-  if(!RLOAD)
-    return krr("main thread only");
-  if (ROPEN < 0)
-    ropen(NULL);
-  ParseStatus status;
-  SEXP txt, sym, val;
-  char rerr[256];extern char	R_ParseErrorMsg[256];
-  char *name = getkstring(x);
-  /* generate symbol to check name is valid */
-  PROTECT(txt=allocVector(STRSXP, 1));
-  SET_STRING_ELT(txt, 0, mkChar(name));
-  free(name);
-  PROTECT(sym = R_ParseVector(txt, 1, &status,R_NilValue));
-  if (status != PARSE_OK) {
-    UNPROTECT(2);
-    snprintf(rerr,sizeof(rerr),"%s: %s",ParseError[status], R_ParseErrorMsg);
-    return krr(rerr);
-  }
-  if(SYMSXP != TYPEOF(VECTOR_ELT(sym,0))){
-    UNPROTECT(2);
-    return krr("nyi");
-  }
-  /* read back symbol string */
-  const char *c = CHAR(PRINTNAME(VECTOR_ELT(sym,0)));
-  PROTECT(val = from_any_kobject(y));
-  defineVar(install(c),val,R_GlobalEnv);
-  UNPROTECT(3);
-  R_ProcessEvents();
-  return (K)0;
-}
-
-__attribute__((constructor)) V __attach(V) {RLOAD=1;}
